Mostly completed use case. Validation, AWS Deployment, and duplicate exceptions not yet implemented.

+++ src/main/java/com/capgemini/mprs/payout/PayoutRepository.java
@@ -2,5 +2,5 @@ package com.capgemini.mprs.payout;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PayoutRepository extends JpaRepository<Payout,Integer> {
+public interface PayoutRepository extends JpaRepository<Payout,Long> {
 }
